Update building info for OSX users
As this project uses `cgo`, OSX users need cross compilation to target
AWS Lambda Linux OS. The only viable solution so far seems to be the use
of docker containers.

`make dbuild` launch a docker container with a Linux OS which in turn
uses the `build` target of the Makefile inside the container to make the
actual build.

CAT: #doc
SEE: #2