Make Lambda function and handler names dynamic
Until now one need to name the output binary "handler.so". The inner
Lambda function had to be "handler" and the Lambda handler had to be
"handle".
With this evolution, we can use gcc preprocessor and set custom
names for both function and handler.

Before:

    go build -buildmode=c-shared -ldflags="-w -s" -o handler.so
    zip handler.zip handler.so

After:

    CGO_FLAGS='-DFUNCTION=index -DHANDLER=custom' \
    go build -buildmode=c-shared -ldflags="-w -s" -o index.so
    zip handler.zip index.so

If any definition is omitted, then it defaults to the current values
"handler" and "handle".

CAT: #feat
RES: #5